finish select with conditions(=, <, >, <>)

+++ api/api.cpp
@@ -9,7 +9,7 @@
 #include <dirent.h>
 #include <sys/stat.h>
 
-API::API(void)
+API::API()
 {
 	this->database_name.clear();
 	this->m_catalog = new Catalog_Manager(this->database_name);
@@ -21,35 +21,37 @@ int API::execute(Statement * s)
 {
 	if (!s)
 		return -1;
+	int ret = 1;
 	try {
-		switch (s->type())
-		{
-			case QUIT_TYPE:
-				std::cout << "Bye" << std::endl;
-				return -1;
-			case USE_TYPE:
-				this->execute_use_database(s);
-				break;
-			case CREATE_DB_TYPE:
-				this->execute_create_database(s);
-				break;
-			case DROP_DB_TYPE:
-				this->execute_drop_database(s);
-				break;
-			case SELECT_TYPE:
-				this->execute_select(s);
-				break;
-			case CREATE_TABLE_TYPE:
-				this->execute_create_table(s);
-				break;
-			case DROP_TABLE_TYPE:
-				this->execute_drop_table(s);
-				break;
-			case INSERT_TYPE:
-				this->execute_insert(s);
-				break;
-			default:
-				break;
+		if (s->type() == QUIT_TYPE) {
+			std::cout << "Bye" << std::endl;
+			ret = -1;
+		} else if (s->type() == USE_TYPE) {
+			this->execute_use_database(s);
+		} else if (s->type() == CREATE_DB_TYPE) {
+			this->execute_create_database(s);
+		} else if (s->type() == DROP_DB_TYPE) {
+			this->execute_drop_database(s);
+		} else {
+			if (this->database_name.empty()) {
+				throw Error(7, "Please use database first.");
+			}
+			switch (s->type()) {
+				case SELECT_TYPE:
+					this->execute_select(s);
+					break;
+				case CREATE_TABLE_TYPE:
+					this->execute_create_table(s);
+					break;
+				case DROP_TABLE_TYPE:
+					this->execute_drop_table(s);
+					break;
+				case INSERT_TYPE:
+					this->execute_insert(s);
+					break;
+				default:
+					break;
+			}
 		}
 		delete s;
 	} catch (Error & e) {
@@ -57,7 +59,7 @@ int API::execute(Statement * s)
 	}
 
 	std::cout << std::endl;
-	return 1;
+	return ret;
 }
 
 void API::execute_use_database(Statement * s)
@@ -84,8 +86,7 @@ void API::execute_create_database(Statement * s)
 void API::execute_drop_database(Statement * s)
 {
 	std::string database_name = std::get<std::string>(s->args());
-	if (opendir(database_name.c_str()))
-	{
+	if (opendir(database_name.c_str())) {
 		if (database_name == this->database_name)
 		{
 			this->database_name.clear();
@@ -94,6 +95,8 @@ void API::execute_drop_database(Statement * s)
 		}
 		std::string cmd = "rm -rf " + database_name;
 		system(cmd.c_str());
+	} else {
+		throw Error(1, "No database named '" + database_name + "'.");
 	}
 }
 
@@ -104,18 +107,34 @@ void API::execute_select(Statement * s)
 	ExpressionList * select_cond = std::get<ExpressionList *>(s->args(2));
 	std::string table_name = std::get<std::string>(((*table_list)[0])->values());
 	std::vector<std::string> list;
+	std::map<std::string, std::pair<int, DMType>> cond;
 	for (int i = 0; i < select_list->size(); ++i) {
 		std::string str = std::get<std::string>((*select_list)[i]->values());
 		list.push_back(str);
 	}
 	if (select_cond) {
 		for (size_t i = 0; i < select_cond->size(); i++) {
-			std::cout << std::get<std::string>(((*select_cond)[i])->values()) << std::endl;
-
+			std::string str = std::get<std::string>(((*select_cond)[i])->values(0));
+			cond[str] = std::make_pair(std::get<int>(((*select_cond)[i])->values(1)), ((*select_cond)[i])->values(2));
 		}
 	}
 	if (this->m_catalog->has_table(table_name)) {
-		this->m_record->select(table_name, list);
+		std::map<std::string, AttrType> & attribute_list = this->m_catalog->get_attributes(table_name);
+		for (int i = 0; i < list.size(); ++i) {
+			if (list[i] == "*") {
+				list.erase(list.begin() + i);
+				for (MapIterator<std::string, AttrType> it = attribute_list.begin(); it != attribute_list.end(); ++it) {
+					list.insert(list.begin() + i, it->first);
+				}
+			}
+			if (attribute_list.find(list[i]) == attribute_list.end())
+				throw Error(701, "Table '" + table_name + "' has no attribute '" + list[i] + "'.");
+		}
+		for (auto & attr : cond) {
+			if (attribute_list.find(attr.first) == attribute_list.end())
+				throw Error(701, "Table '" + table_name + "' has no attribute '" + attr.first + "'.");
+		}
+		this->m_record->select(table_name, list, cond);
 	} else {
 		throw Error(700, "No table named '" + table_name + "'.");
 	}
@@ -128,13 +147,15 @@ void API::execute_insert(Statement * s)
 	ExpressionList * attribute_list = std::get<ExpressionList *>(s->args(1));
 	ExpressionList * value_list = std::get<ExpressionList *>(s->args(2));
 	std::map<std::string, DMType> attr_value;
+	if (!this->m_catalog->has_table(table_name)) {
+		throw Error(700, "No table named '" + table_name + "'.");
+	}
 	if (attribute_list->size() != value_list->size())
-		return;
+		throw Error(756, "The values are incompleted.");
 	for (size_t i = 0; i < attribute_list->size(); i++)
 	{
 		std::string attr = std::get<std::string>((*attribute_list)[i]->values());
 		attr_value[attr] = (*value_list)[i]->values();
-//		std::cout << attr << std::endl;
 	}
 
 	this->m_record->insert(table_name, attr_value);
@@ -163,12 +184,14 @@ void API::execute_create_table(Statement * s)
 void API::execute_drop_table(Statement * s)
 {
 	std::string table_name = std::get<std::string>(s->args());
+	if (!this->m_catalog->has_table(table_name)) {
+		throw Error(700, "No table named '" + table_name + "'.");
+	}
 	this->m_catalog->drop_table(table_name);
 }
 
-API::~API(void)
+API::~API()
 {
-//	std::cout << "~API" << std::endl;
 	delete this->m_catalog;
 	delete this->m_buffer;
 	delete this->m_record;
